Final Form

+++ PharmaceuticalWarehouseManagementSystem/PharmaceuticalWarehouseManagementSystem.UI/Areas/User/Controllers/CustomerController.cs
@@ -14,14 +14,15 @@ namespace PharmaceuticalWarehouseManagementSystem.UI.Areas.User.Controllers
     [Area("User")]
     [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme)]
     [Authorize(Roles = "User")]
-        public class CustomerController : Controller
+         public class CustomerController : Controller
     {
-        private ICustomerRepository _repository;
-        private ILogger<CustomerController> _logger;
+        private readonly ICustomerRepository _repository;
+        private readonly ILogger<CustomerController> _logger;
 
-        public CustomerController(ICustomerRepository repository)
+        public CustomerController(ICustomerRepository repository,ILogger<CustomerController> logger)
         {
             this._repository = repository;
+            this._logger = logger;
         }
 
         
@@ -49,20 +50,20 @@ namespace PharmaceuticalWarehouseManagementSystem.UI.Areas.User.Controllers
                 if (result == true)
                 {
                     _repository.Save();
-                    _logger.LogWarning("Customer Added"+" "+DateTime.Now.ToString());
+                    _logger.LogInformation("Customer Added"+" "+DateTime.Now.ToString());
                     return RedirectToAction("List","Customer");
                 }
                 else
                 {
                     TempData["Message"] = $"Kayıt işlemi sırasında bir hata oluştu. Lütfen tüm alanları kontrol edip tekrar deneyin..!";
-                    _logger.LogWarning("Customer adding operation failed"+" "+ DateTime.Now.ToString());
+                    _logger.LogError("Customer adding operation failed"+" "+ DateTime.Now.ToString());
                     return View(customer);
                 }
             }
             else
             {
                 TempData["Message"] = $"Kayıt işlemi sırasında bir hata oluştu. Lütfen tüm alanları kontrol edip tekrar deneyin..!";
-                _logger.LogError("Customer Add operations Error" +" "+DateTime.Now.ToString());
+                _logger.LogCritical("Customer Add operations Error" +" "+DateTime.Now.ToString());
                 return View(customer);
             }
         }
@@ -103,14 +104,14 @@ namespace PharmaceuticalWarehouseManagementSystem.UI.Areas.User.Controllers
                 else
                 {
                     TempData["Message"] = $"Güncelleme işlemi sırasında bir hata oluştu. Lütfen tekrar deneyin..!";
-                    _logger.LogWarning("Customer edit opertaion failed"+" "+DateTime.Now.ToString());
+                    _logger.LogError("Customer edit opertaion failed"+" "+DateTime.Now.ToString());
                     return View(updated);
                 }
             }
             else
             {
                 TempData["Message"] = $"Güncelleme işlemi sırasında bir hata oluştu. Lütfen tekrar deneyin..!";
-                  _logger.LogError("Customer edit opertaion failed"+" "+DateTime.Now.ToString());
+                  _logger.LogCritical("Customer edit opertaion failed"+" "+DateTime.Now.ToString());
                 return View();
             }
         }
@@ -121,9 +122,17 @@ namespace PharmaceuticalWarehouseManagementSystem.UI.Areas.User.Controllers
 
         public IActionResult Delete(Guid id)
         {
-             _logger.LogInformation("Customer Deleted"+" "+DateTime.Now.ToString());
-            _repository.Remove(_repository.GetById(id));
-            return RedirectToAction("List");
+            if (ModelState.IsValid)
+            {
+                _logger.LogInformation("Customer Deleted"+" "+ id+" "+DateTime.Now.ToString());
+                _repository.Remove(_repository.GetById(id));
+                return RedirectToAction("List");
+            }
+            else
+            {
+                _logger.LogError("Customer Delete Action Failed"+" "+DateTime.Now.ToString());
+                return BadRequest();
+            }
         }
